Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -8,7 +8,7 @@ int main() {
     scanf("%d",&y);
     scanf("%d",&z);
 
-    int avg = x+y+z/3;
+    float avg = x+y+z/3;
     printf("Average: %.2f",sum);
     return 0;
 }+
